GPG-517: Markups-Removed "New" from Classes, functions & files
- Changed Late Submission spacing
- Refactored reporting years helper
- Removed Unused view models
