+++ GenderPayGap.WebUI/Models/Report/LateSubmissionReasonViewModel.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Mvc.ModelBinding;
 
 namespace GenderPayGap.WebUI.Models.Report
 {
-    public class NewLateSubmissionReasonViewModel : GovUkViewModel
+    public class LateSubmissionReasonViewModel : GovUkViewModel
     {
 
         [BindNever /* Output Only - only used for sending data from the Controller to the View */]
@@ -21,13 +21,13 @@ namespace GenderPayGap.WebUI.Models.Report
         public DateTime DeadlineDate { get; set; }
 
         [GovUkValidateRequired(ErrorMessageIfMissing = "Select yes if you have received a letter about a breach of the regulations from the Equality and Human Rights Commission.")]
-        public NewReportLateSubmissionReceivedLetterFromEhrc? ReceivedLetterFromEhrc { get; set; }
+        public ReportLateSubmissionReceivedLetterFromEhrc? ReceivedLetterFromEhrc { get; set; }
         [GovUkValidateRequired(ErrorMessageIfMissing = "Explain why your organisation is reporting or changing your gender pay gap figures after the deadline.")]
         public string Reason { get; set; }
 
     }
 
-    public enum NewReportLateSubmissionReceivedLetterFromEhrc
+    public enum ReportLateSubmissionReceivedLetterFromEhrc
     {
         Yes,
         No
