+++ GenderPayGap.WebUI/Models/Report/ReportFiguresViewModel.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Mvc.ModelBinding;
 
 namespace GenderPayGap.WebUI.Models.Report
 {
-    public class NewReportFiguresViewModel : GovUkViewModel
+    public class ReportFiguresViewModel : GovUkViewModel
     {
 
         [BindNever /* Output Only - only used for sending data from the Controller to the View */]
