+++ GenderPayGap.WebUI/Views/ReportFigures/ReportFigures.cshtml
@@ -2,7 +2,7 @@
 @using GenderPayGap.WebUI.Views.Components.StatusMessages
 @using GovUkDesignSystem
 @using GovUkDesignSystem.GovUkDesignSystemComponents
-@model GenderPayGap.WebUI.Models.Report.NewReportFiguresViewModel
+@model GenderPayGap.WebUI.Models.Report.ReportFiguresViewModel
 
 @{
     string encryptedOrganisationId = Encryption.EncryptQuerystring(Model.Organisation.OrganisationId.ToString());
@@ -33,7 +33,7 @@
             new CrumbViewModel
             {
                 Text = $"{editOrReportTitle} (reporting year {formattedReportingYears})",
-                Href = Url.Action("NewReportOverview", "NewReportOverview", new {encryptedOrganisationId = encryptedOrganisationId, reportingYear = Model.ReportingYear})
+                Href = Url.Action("ReportOverview", "ReportOverview", new {encryptedOrganisationId = encryptedOrganisationId, reportingYear = Model.ReportingYear})
             },
             new CrumbViewModel
             {
@@ -48,7 +48,7 @@
     }))
 }
 
-<form method="post" action="@(Url.Action("NewReportFiguresPost", "NewReportFigures", new { encryptedOrganisationId = encryptedOrganisationId, reportingYear = Model.ReportingYear }))">
+<form method="post" action="@Url.Action("ReportFiguresPost", "ReportFigures", new { encryptedOrganisationId = encryptedOrganisationId, reportingYear = Model.ReportingYear })">
 
     <div class="govuk-grid-row">
         <div class="govuk-grid-column-two-thirds">
@@ -312,12 +312,12 @@
             }))
 
             <p class="govuk-body">
-                <a class="govuk-link" href="@Url.Action("NewReportOverview", "NewReportOverview", new {encryptedOrganisationId = encryptedOrganisationId, reportingYear = Model.ReportingYear})">
+                <a class="govuk-link" href="@Url.Action("ReportOverview", "ReportOverview", new {encryptedOrganisationId = encryptedOrganisationId, reportingYear = Model.ReportingYear})">
                     Cancel and return to report overview
                 </a>
             </p>
         </div>
-        @await Html.PartialAsync("~/Views/Submit/NewRelatedContent.cshtml")
+        @await Html.PartialAsync("~/Views/Submit/RelatedContent.cshtml")
     </div>
 
 </form>
