+++ GenderPayGap.WebUI/Views/Viewing/EmployerDetails/Employer.cshtml
@@ -185,7 +185,7 @@
 
             </div>
 
-            @await Html.PartialAsync("/Views/Submit/NewRelatedContent.cshtml")
+            @await Html.PartialAsync("/Views/Submit/RelatedContent.cshtml")
         </div>
     </div>
 </div>
